ULongs can map to a java long, not BigInteger.  Thus, avoid creating BigInteger objects if we can.
git-svn-id: https://svn.apache.org/repos/asf/cxf/trunk@683391 13f79535-47bb-0310-9956-ffa450edef68

+++ rt/bindings/corba/src/main/java/org/apache/cxf/binding/corba/runtime/CorbaObjectReader.java
@@ -212,11 +212,11 @@ public class CorbaObjectReader {
         }
     }
 
-    public BigInteger readULong() throws CorbaBindingException {
+    public long readULong() throws CorbaBindingException {
         try {
             long l = stream.read_ulong();
             l &= 0xffffffffL;
-            return BigInteger.valueOf(l);
+            return l;
         } catch (org.omg.CORBA.MARSHAL ex) {
             LOG.log(Level.SEVERE, "CorbaObjectReader: could not read unsigned long");
             throw new CorbaBindingException("CorbaObjectReader: readULong MARSHAL exception", ex);
